+++ imports/lib/models/settings.ts
@@ -2,11 +2,11 @@ import { check, Match } from 'meteor/check';
 import { Meteor } from 'meteor/meteor';
 import { Mongo } from 'meteor/mongo';
 import isAdmin from '../is-admin';
-import SettingsSchema, { SettingType } from '../schemas/settings';
+import SettingSchema, { SettingType } from '../schemas/setting';
 import Base, { FindOptions } from './base';
 
 const Settings = new Base<SettingType>('settings');
-Settings.attachSchema(SettingsSchema);
+Settings.attachSchema(SettingSchema);
 
 // Publish manually instead of through Base.publish because we need to block the
 // query for non-admins, and Base.publish doesn't support permission checks.
