+++ imports/server/hunts.ts
@@ -15,7 +15,7 @@ import {
   userMayUseDiscordBotAPIs,
 } from '../lib/permission_stubs';
 import { HuntType } from '../lib/schemas/hunt';
-import { SettingType } from '../lib/schemas/settings';
+import { SettingType } from '../lib/schemas/setting';
 import addUserToDiscordRole from './addUserToDiscordRole';
 import List from './blanche';
